+++ src/config/config_manager.h
@@ -60,7 +60,7 @@ public:
     ConfigManager(fs::path filename,
         const fs::path& userhome, const fs::path& config_dir,
         fs::path prefix_dir, fs::path magic_file,
-        std::string ip, std::string interface, int port,
+        std::string ip, std::string interface, unsigned short port,
         bool debug_logging);
     virtual ~ConfigManager() override;
 
@@ -147,7 +147,7 @@ protected:
     fs::path magic_file;
     std::string ip;
     std::string interface;
-    int port;
+    unsigned short port;
     static bool debug_logging;
     std::map<std::string, std::string> origValues;
